refactor: move each function in sub-folder
