+++ src/types/constant.ts
@@ -1,0 +1,3 @@
+export const STATUS_BAR_SELECTOR = "body > div.app-container div.status-bar";
+export const RIBBON_BAR_SELECTOR = "body > div.app-container div.side-dock-actions";
+export const ANIMATION_DURATION = 500;
