+++ src/utils.ts
@@ -1,6 +1,3 @@
-import Fuse from "fuse.js";
-import { type App, type ChildElement, requireApiVersion } from "obsidian";
-
 export function getPreviousSiblings(
 	el: HTMLElement,
 	filter?: (el: HTMLElement) => boolean
@@ -76,107 +73,3 @@ export function reorderArray(array: any[], from: number, to: number, on = 1) {
 }
 
 // flatten infinity scroll root elements
-
-export const getItems = (items: ChildElement[], app: App): ChildElement[] => {
-	let children: any[] = [];
-	const supportsVirtualChildren = requireApiVersion && requireApiVersion("0.15.0");
-	const excluded = app.vault.config.userIgnoreFilters;
-	if (excluded) {
-		items = items.filter((item) => {
-			//if the item.file.path startswith any of the excluded paths, return false
-			return !excluded.some((exclude) => item.file.path.startsWith(exclude));
-		});
-	}
-	return supportsVirtualChildren
-		? items
-				.reduce((res, item) => {
-					if (item.vChildren?._children) {
-						children = [...children, ...item.vChildren._children];
-					} else {
-						res.push(item);
-					}
-					return res;
-				}, [] as ChildElement[])
-				.concat(children.length ? getItems(children, app) : children)
-		: items
-				.reduce((res, item) => {
-					if (item.children) {
-						children = [...children, ...item.children];
-					} else {
-						res.push(item);
-					}
-					return res;
-				}, [] as ChildElement[])
-				.concat(children.length ? getItems(children, app) : children);
-};
-
-// highlight fuzzy filter matches
-
-type NestedObject = { [key: string]: string | NestedObject };
-
-export const highlight = (
-	fuseSearchResult: any,
-	highlightClassName: string = "suggestion-highlight"
-) => {
-	const set = (obj: NestedObject, path: string, value: any) => {
-		const pathValue = path.split(".");
-		let i;
-
-		for (i = 0; i < pathValue.length - 1; i++) {
-			obj = obj[pathValue[i]] as NestedObject;
-		}
-
-		obj[pathValue[i]] = value;
-	};
-
-	const generateHighlightedText = (inputText: string, regions: number[][] = []) => {
-		return regions
-			.reduce((str, [start, end]) => {
-				str[start] = `<span class="${highlightClassName}">${str[start]}`;
-				str[end] = `${str[end]}</span>`;
-				return str;
-			}, inputText.split(""))
-			.join("");
-	};
-
-	return fuseSearchResult
-		.filter(({ matches }: any) => matches && matches.length)
-		.map(({ item, matches }: any) => {
-			const highlightedItem = { ...item };
-			matches.forEach((match: any) => {
-				if (!highlightedItem.innerEl.origContent)
-					highlightedItem.innerEl.origContent = highlightedItem.innerEl.textContent;
-				set(
-					highlightedItem,
-					"innerEl.innerHTML",
-					generateHighlightedText(match.value, match.indices)
-				);
-				highlightedItem.innerEl?.addClass("has-matches");
-			});
-
-			return highlightedItem;
-		});
-};
-
-export function removeExt(obj: any) {
-	if (typeof obj === "string" || obj instanceof String) {
-		const parts = obj.split("/");
-		let newObj = obj;
-		if (parts.length >= 3) {
-			for (let i = 1; i < parts.length - 1; i += 2) {
-				parts[i] = "…";
-			}
-			newObj = parts.join("/");
-		}
-		return newObj.replace(/.md$/, "");
-	}
-	return obj;
-}
-
-export function getFn(obj: any, path: string[]) {
-	const value = Fuse.config.getFn(obj, path);
-	if (Array.isArray(value)) {
-		return value.map((el) => removeExt(el));
-	}
-	return removeExt(value);
-}
